+++ CityLBM/src/Components/Results/SliceVisualizationComponent.cs
@@ -343,7 +343,7 @@ namespace CityLBM.Components.Results
 
         protected override Bitmap Icon
         {
-            get { return null; }
+            get { return null!; }
         }
 
         public override Guid ComponentGuid